+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Move the prime field module from `crypto/` to `types/`[PR 602](https://github.com/functionalscript/functionalscript/pull/602)
+- Digital signatures [PR 599](https://github.com/functionalscript/functionalscript/pull/599)
+
 ## 0.8.1
 
 - 64bit SHA2 padding is fixed [PR 595](https://github.com/functionalscript/functionalscript/pull/595)